added alignment score filtering; now supporting restarts for disrupted runs

+++ batch.py
@@ -30,7 +30,7 @@ def count_file_lines(path):
     return int(wc_output.split()[0])
 
 
-def do_map(fastq1, fastq2, refpath, bowtie_threads, min_match_len, min_mapq):
+def do_map(fastq1, fastq2, refpath, bowtie_threads, min_match_len, min_mapq, min_score):
     """
     Process SAM output as it is streamed from bowtie2 to assign
     short reads to HCV genotypes/subtypes.
@@ -55,6 +55,7 @@ def do_map(fastq1, fastq2, refpath, bowtie_threads, min_match_len, min_mapq):
     n_short = 0
     n_hybrid = 0
     n_mapq = 0
+    n_low = 0
     progress = 0
     total_count = 0
     
@@ -63,14 +64,16 @@ def do_map(fastq1, fastq2, refpath, bowtie_threads, min_match_len, min_mapq):
             if line.startswith('@'):
                 # skip header line
                 continue
-            
-            qname, flag, rname, pos, mapq, cigar, rnext, pnext, tlen, seq, qual = line.split('\t')[:11]
+
+            items = line.split('\t')
+            qname, flag, rname, pos, mapq, cigar, rnext, pnext, tlen, seq, qual = items[:11]
             progress += 1
             if progress % 5000 == 0:
-                print('[%s] (%d/%d) mapped %d' % (datetime.now().strftime('%Y-%m-%d %H:%M:%S'),
-                    progress, nrecords, 2*total_count))
+                print('[%s] (%d/%d) mapped %d (%d/%d/%d/%d)' % (datetime.now().strftime('%Y-%m-%d %H:%M:%S'),
+                    progress, nrecords, 2*total_count, n_mapq, n_hybrid, n_short, n_low))
                 
             subtype = rname.split('-')[1]  # 'HCV-1a' -> '1a'
+            genotype = subtype[0]
             
             # ignore second reads
             if not is_first_read(flag):
@@ -81,18 +84,23 @@ def do_map(fastq1, fastq2, refpath, bowtie_threads, min_match_len, min_mapq):
             if q < min_mapq:
                 n_mapq += 1
                 continue
-            
-            # ignore reads whose mate mapped to a different reference
-            if rnext != '=':
-                #print rname, rnext
+
+            # ignore reads whose mate mapped to a different genotype
+            if rnext != '=' and genotype != rnext.split('-')[1][0]:
                 n_hybrid += 1
                 continue
-                
-            # filter out primers based on match length
+            
+            # filter out reads based on match length - e.g., primers
             tokens = cigar_re.findall(cigar)
             match_len = sum([int(token.strip('M')) for token in tokens if token.endswith('M')])
             if match_len < min_match_len:
                 n_short += 1
+                continue
+
+            # filter out reads with low alignment score
+            optionals = dict([item.split(':i:') for item in items[11:] if ':i:' in item])
+            if 'AS' not in optionals or int(optionals['AS']) < min_score:
+                n_low += 1
                 continue
 
             # update counts
@@ -116,16 +124,29 @@ def main():
     parser = argparse.ArgumentParser(
         description='Map contents of FASTQ R1 and R2 data sets to references using bowtie2.')
     
-    parser.add_argument('path', nargs='?', default=sys.stdin.read().splitlines(), 
-                            help='<input> folder containing FASTQ data')
+    parser.add_argument('-path', help='<input> folder containing FASTQ files')
+    parser.add_argument('-pathlist', help='<input> file containing list of folders')
     parser.add_argument('-output', help='<output CSV> file to write results')
     parser.add_argument('-log', help='path to logfile', default='mixed-hcv.batch.log')
     parser.add_argument('-x', help='path to bowtie2 index (.bt2)', default='data/gb-ref')
     parser.add_argument('-p', type=int, help='number of bowtie2 threads', default=6)
-    parser.add_argument('-minlen', type=int, help='minimum match length (CIGAR M)', default=50)
+    parser.add_argument('-minlen', type=int, help='minimum match length (CIGAR M)', default=200)
     parser.add_argument('-minq', type=int, help='minimum mapping quality (MAPQ)', default=0)
+    parser.add_argument('-mins', type=int, help='minimum alignment score', default=200)
 
     args = parser.parse_args()
+    
+    paths = []
+    if args.pathlist:
+        handle = open(args.pathlist, 'rU')
+        for line in handle:
+            paths.append(line.strip('\n'))
+        handle.close()
+    elif args.path:
+        paths.append(args.path)
+    else:
+        parser.print_help()
+        sys.exit()
     
     # check that we have access to bowtie2
     try:
@@ -133,12 +154,24 @@ def main():
     except OSError:
         raise RuntimeError('bowtie2 not found; check if it is installed and in $PATH\n')
     
-    logfile = open(args.log, 'a')
     
-    handle = open(args.output, 'w')
-    handle.write('runname,sample,snum,subtype,count,total\n')
     
-    for path in args.path:
+    complete = {}
+    if os.path.exists(args.output):
+        handle = open(args.output, 'rU')
+        header = handle.next()
+        for line in handle:
+            runname, sample, snum = line.strip('\n').split(',')[:3]
+            complete.update({(runname, sample, snum): 1})
+        handle.close()
+        # re-open file for appending
+        handle = open(args.output, 'a')
+    else:
+        # new output file
+        handle = open(args.output, 'w')
+        handle.write('runname,sample,snum,subtype,count,total\n')
+    
+    for path in paths:
         # check that the inputs exist
         if not os.path.exists(path):
             print('No folder found at %s', path)
@@ -159,13 +192,18 @@ def main():
             filename = os.path.basename(f1)
             if filename.startswith('Undetermined'):
                 continue
-                
-            logfile.write('[%s] start processing %s\n' % (datetime.now().strftime('%Y-%m-%d %H:%M:%S'), filename))
-            logfile.flush()
             
             sample, snum = filename.split('_')[:2]
+            
+            if (runname, sample, snum) in complete:
+                # skip previously processed sample
+                continue
+                
+            logfile = open(args.log, 'a')
+            logfile.write('[%s] start processing %s\n' % (datetime.now().strftime('%Y-%m-%d %H:%M:%S'), filename))
+            logfile.close()
 
-            counts, n_short, n_mapq, n_hybrid = do_map(f1, f2, args.x, args.p, args.minlen, args.minq)
+            counts, n_short, n_mapq, n_hybrid = do_map(f1, f2, args.x, args.p, args.minlen, args.minq, args.mins)
             n_discard = n_short + n_mapq + n_hybrid
             total_count = sum(counts.values()) + n_discard
 
@@ -176,8 +214,9 @@ def main():
             handle.write('%s,%s,%s,,%d,%d\n' % (runname, sample, snum, n_discard, total_count))
             handle.flush()  # clear write buffer
             
+            logfile = open(args.log, 'a')
             logfile.write('[%s] end processing %s\n' % (datetime.now().strftime('%Y-%m-%d %H:%M:%S'), filename))
-            logfile.flush()
+            logfile.close()
             
     handle.close()
     logfile.close()
